+++ common/src/python/rxnav/rxclass.py
@@ -1,0 +1,8 @@
+"""
+List of RxClasses of interest.
+"""
+
+ALL_RXCLASSES = [
+	"C02L",
+	"... TODO FILL OUT REST"
+]