+++ common/test/python/rxnorm/test_rxnorm.py
@@ -1,15 +1,15 @@
-"""Tests the RxNorm API, which is public so doesn't need any authorization."""
+"""Tests the RxNav API, which is public so doesn't need any authorization."""
 
-from rxnorm.rxnorm_connection import RxcuiStatus, RxNormConnection
+from RxNav.RxNav_connection import RxcuiStatus, RxNavConnection
 
 
-class TestRxNormConnection:
-    """Tests the RxNormConnection class."""
+class TestRxNavConnection:
+    """Tests the RxNavConnection class."""
 
     def test_url_creation(self):
         """Test URL creation."""
         assert (
-            RxNormConnection.url("REST/test/path")
+            RxNavConnection.url("REST/test/path")
             == "https://rxnav.nlm.nih.gov/REST/test/path"
         )
 
@@ -18,9 +18,9 @@ class TestRxNormConnection:
         Test the get_rxcui_status method - uses same examples defined on
         https://lhncbc.nlm.nih.gov/RxNav/APIs/api-RxNorm.getRxcuiHistoryStatus.html
         """
-        assert RxNormConnection.get_rxcui_status(1801289) == RxcuiStatus.ACTIVE
-        assert RxNormConnection.get_rxcui_status(861765) == RxcuiStatus.OBSOLETE
-        assert RxNormConnection.get_rxcui_status(105048) == RxcuiStatus.REMAPPED
-        assert RxNormConnection.get_rxcui_status(1360201) == RxcuiStatus.QUANTIFIED
-        assert RxNormConnection.get_rxcui_status(3686) == RxcuiStatus.NOT_CURRENT
-        assert RxNormConnection.get_rxcui_status(0) == RxcuiStatus.UNKNOWN
+        assert RxNavConnection.get_rxcui_status(1801289) == RxcuiStatus.ACTIVE
+        assert RxNavConnection.get_rxcui_status(861765) == RxcuiStatus.OBSOLETE
+        assert RxNavConnection.get_rxcui_status(105048) == RxcuiStatus.REMAPPED
+        assert RxNavConnection.get_rxcui_status(1360201) == RxcuiStatus.QUANTIFIED
+        assert RxNavConnection.get_rxcui_status(3686) == RxcuiStatus.NOT_CURRENT
+        assert RxNavConnection.get_rxcui_status(0) == RxcuiStatus.UNKNOWN