+++ docs/attribute_curator/CHANGELOG.md
@@ -9,8 +9,9 @@ All notable changes to this gear are documented in this file.
     * Updates to support curating over dicom and nifti files, and expands the `filename_pattern` to `filename_patterns` to support multiple patterns
     * Adds `img_study_date` to use the `file.info.header.dicom.StudyDate` field for image file ordering
     * Adds `file.info._filename` as temporary metadata added to the table for deriver use (needed for curation of images file locators)
+* Queries RxClass API for V4+ A4 derived variable calculations, and passes to the deriver through the table under `_rxclass`
 * Adds loop for missingness logic
-    * Passes the previous record to the deriver to support missingness needs under `file.info._prev_recors`
+    * Passes the previous record to the deriver to support missingness needs through the table under `_prev_record`
 * Updates function names from `pre/post_process` to `pre/post_curate`
 * Parameterizes number of workers